fix: resolve remaining deterministic and SQL parsing review findings
Address CodeRabbit findings by broadening SQL column type parsing, replacing locale-sensitive sort paths, and centralizing deterministic string sort helpers for workflow modules.
