+++ src/schema/erd-extractor.js
@@ -24,7 +24,8 @@ const SQL_TABLE_FOREIGN_KEY_RE = new RegExp(
   'i'
 );
 const SQL_TABLE_CONSTRAINT_LINE_RE = /^(?:constraint|foreign\s+key|primary\s+key|unique)\b/i;
-const SQL_COLUMN_DEFINITION_RE = /^([`"]?[A-Za-z_][A-Za-z0-9_]*[`"]?)\s+([A-Za-z0-9_()]+)([\s\S]*)$/;
+const SQL_COLUMN_DEFINITION_RE =
+  /^([`"]?[A-Za-z_][A-Za-z0-9_]*[`"]?)\s+(.+?)(?=\s+(?:constraint\b|not\b|null\b|default\b|references\b|primary\b|unique\b|check\b|generated\b)|$)([\s\S]*)$/i;
 const SCHEMA_PARSERS = Object.freeze({
   prisma: parsePrismaSchema,
   sql: parseSqlSchema,
@@ -219,7 +220,7 @@ function parseSqlSchema(fileContent) {
       const columnMatch = line.match(SQL_COLUMN_DEFINITION_RE);
       if (!columnMatch) continue;
       const columnName = tableNameFromSql(columnMatch[1]);
-      const columnType = columnMatch[2];
+      const columnType = String(columnMatch[2] || '').trim();
       const remainder = columnMatch[3] || '';
       const remainderLower = remainder.toLowerCase();
       const keyFlags = [];
